Added remaining sections in the Stats Section + adjusted Auth to depend on User instead of Token

+++ app/src/main/java/com/example/shikiflow/data/mapper/common/DateMapper.kt
@@ -43,5 +43,5 @@ object DateMapper {
     }
 
 
-    fun Int.minutesToDays(): Double = this / 60.0 / 24.0
+    fun Int.minutesToDays(): Float = this / 60.0f / 24.0f
 }
